+++ elastic-job-cloud/elastic-job-cloud-scheduler/src/main/java/com/dangdang/ddframe/job/cloud/scheduler/config/app/CloudAppConfigurationService.java
@@ -84,7 +84,6 @@ public final class CloudAppConfigurationService {
         return result;
     }
     
-    
     /**
      * 删除云作业App配置.
      *