+++ elastic-job-lite/elastic-job-lite-lifecycle/src/test/java/com/dangdang/ddframe/job/lite/lifecycle/internal/statistics/JobStatisticsAPIImplTest.java
@@ -54,7 +54,7 @@ public final class JobStatisticsAPIImplTest {
     }
     
     @Test
-    public void assertGetJobBriefInfo() {
+    public void assertGetOKJobBriefInfo() {
         when(regCenter.getChildrenKeys("/")).thenReturn(Lists.newArrayList("test_job"));
         when(regCenter.get("/test_job/config")).thenReturn(LifecycleJsonConstants.getSimpleJobJson("test_job", "desc"));
         when(regCenter.getChildrenKeys("/test_job/servers")).thenReturn(Arrays.asList("ip1", "ip2"));
@@ -74,11 +74,26 @@ public final class JobStatisticsAPIImplTest {
     }
     
     @Test
+    public void assertGetOKJobBriefInfoWithPartialDisabledServer() {
+        when(regCenter.getChildrenKeys("/")).thenReturn(Lists.newArrayList("test_job"));
+        when(regCenter.get("/test_job/config")).thenReturn(LifecycleJsonConstants.getSimpleJobJson("test_job", "desc"));
+        when(regCenter.getChildrenKeys("/test_job/servers")).thenReturn(Arrays.asList("ip1", "ip2"));
+        when(regCenter.get("/test_job/servers/ip1")).thenReturn("DISABLED");
+        when(regCenter.getChildrenKeys("/test_job/instances")).thenReturn(Arrays.asList("ip1@-@defaultInstance", "ip2@-@defaultInstance"));
+        when(regCenter.getChildrenKeys("/test_job/sharding")).thenReturn(Arrays.asList("0", "1"));
+        when(regCenter.get("/test_job/sharding/0/instance")).thenReturn("ip1@-@defaultInstance");
+        when(regCenter.get("/test_job/sharding/1/instance")).thenReturn("ip2@-@defaultInstance");
+        JobBriefInfo jobBrief = jobStatisticsAPI.getJobBriefInfo("test_job");
+        assertThat(jobBrief.getStatus(), is(JobStatus.OK));
+    }
+    
+    @Test
     public void assertGetDisabledJobBriefInfo() {
         when(regCenter.getChildrenKeys("/")).thenReturn(Lists.newArrayList("test_job"));
         when(regCenter.get("/test_job/config")).thenReturn(LifecycleJsonConstants.getSimpleJobJson("test_job", "desc"));
         when(regCenter.getChildrenKeys("/test_job/servers")).thenReturn(Arrays.asList("ip1", "ip2"));
         when(regCenter.get("/test_job/servers/ip1")).thenReturn("DISABLED");
+        when(regCenter.get("/test_job/servers/ip2")).thenReturn("DISABLED");
         when(regCenter.getChildrenKeys("/test_job/instances")).thenReturn(Arrays.asList("ip1@-@defaultInstance", "ip2@-@defaultInstance"));
         when(regCenter.getChildrenKeys("/test_job/sharding")).thenReturn(Arrays.asList("0", "1"));
         when(regCenter.get("/test_job/sharding/0/instance")).thenReturn("ip1@-@defaultInstance");